% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_response_matrix.R
\name{construct_response_matrix}
\alias{construct_response_matrix}
\title{Construct an item response matrix from ROAR-LDT data}
\usage{
construct_response_matrix(d)
}
\arguments{
\item{d}{A dataframe with one row per respondent. Stimulii should be the
names of columns}
}
\value{
A matrix suitable for factor score estimation using pre-calibrated
mirt model
}
\description{
Extracts a dichotomously coded response matrix to estimate ability scores
using pre-calibrated mirt model. This function is internally called by
estimate_wj() and is unlikely to be particularly useful otherwise
}
