% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_wj.R
\name{estimate_wj}
\alias{estimate_wj}
\title{Estimate Woodcock-Johnson Letter-Word ID scores from ROAR-LDT responses}
\usage{
estimate_wj(d)
}
\arguments{
\item{d}{A dataframe with one row per respondent. Stimulii should be the
names of columns. To estimate WJ scores, must also include a `visit_age`
column where age is measured in months (can be continuously valued).}
}
\value{
A dataframe with columns for subject code, IRT-estimated ROAR-LDT
ability score, estimated WJ-LWID scores (raw and standard), and estimated
WJ-LWID percentile.
}
\description{
Estimates a variety of scores from ROAR-LDT responses. Function will only
consider item responses with precalibrated item parameters, but will handle
missing item responses gracefully. Models to estimate WJ scores depend on the
age of the respondent. If a `visit_age` is not supplied for an individual,
their WJ columns will be populated with NA. Additionally, since all
respondents perform near ceiling and age is factored into the WJ standard
score, estimation is poorly behaved for older respondents.
}
